fix: Fixing product UI issues
