+++ moon-store-mac-os/moon-store-mac-os/Helpers/Modifiers/SpinnerModifier.swift
@@ -17,7 +17,6 @@ struct SpinnerModifier: ViewModifier {
     func body(content: Content) -> some View {
         if showSpinner {
             content
-                .frame(maxSquare: .infinity)
                 .overlay {
                     MSSpinner()
                 }
