+++ moon-store-mac-os/moon-store-mac-os/Views/Product/ListView/ProductListConstants.swift
@@ -1,0 +1,33 @@
+//
+//  ProductListConstants.swift
+//  moon-store-mac-os
+//
+//  Created by Steven Santeliz on 4/1/25.
+//
+
+import Foundation
+
+enum ProductListConstants {
+    static let cornerRadius: CGFloat = 10
+    static let iconSize: CGFloat = 30
+    static let spacing: CGFloat = 15
+    static let plusIcon: String = "plus.circle"
+    static let personIcon: String = "person.circle.fill"
+
+    enum ProductRow {
+        static let horizontalPadding: CGFloat = 5
+        static let padding: CGFloat = 6
+        static let spacing: CGFloat = 10
+        static let pensilIcon: String = "pencil"
+        static let trashIcon: String = "trash"
+        static let optionTitle: String = "Abastecer"
+        static let iconSize: CGFloat = 20
+        static let height: CGFloat = 50
+        static let lineLimit: Int = 1
+        static let opacity: CGFloat = 0.1
+        static let pairNumber: Int = 2
+        static let categorySize: CGFloat = 100
+        static let stockSize: CGFloat = 80
+        static let optionsSise: CGFloat = 150
+    }
+}
